requirements.txt
numpy==1.21.0
Werkzeug==2.0.1
Flask==2.0.1
torch==1.9.0+cpu
Pillow==8.2.0
gunicorn == 20.1.0
pandas==1.2.5
matplotlib==3.4.2
